requirements.txt
av
color-matcher
diffusers
ftfy
gguf
imageio-ffmpeg
insightface
mss
onnxruntime
onnxruntime-gpu
peft
sageattention
scikit-video
segment-anything
sentencepiece
spandrel
toml
torchao
torchaudio
torchsde
ultralytics
xformers
